Recover some VOD titles from 2022
These titles were recovered from twitchtracker.com. This service doesn't
seem completely reliable as it didn't catch all of the streams, it
repeats the title for some streams, and it sometimes doesn't have the
alternate title for a stream that I know was used. I decided to copy
over the titles which are not repeated between consecutive streams for
videos I have failed to record the title of.

Note that the 2022-04-05 stream on twitchtracker.com has an alternate
title which was added later on in the stream which I feel is correct, so
I've added it to the end of the title as I do for other streams where
the title changes mid-stream.