Add streamcharts data
Instead of manually checking streamscharts for missing streams or
missing titles, it would be easier to automate the process.

Unfortunately, streamcharts.com does not offer a free API or a
reasonably-priced tier for API access. Fortunately, the only data I
want are the stream times and their titles. So, I manually scraped the
streams summary view (9 pages in total) using a multiselect cursor in
Sublime to convert the data from HTML to CSV.

Note that the times in the summary view are inaccurate, for some reason,
but will be close enough to find the title for a stream I have. For
streams I am missing, I will have to load the detail view of the stream
to get the actual time.