Write history in chronological order
It makes more sense to write the history in chronological order instead
of reverse order, since that's the same direction the text flows on the
paeg.

Also adjusted the way the history is written. The rendering of the
history section looks odd in Obsidian due to the bullet points, but we
need newlines between each history entry. Otherwise github will render
all of the links on the same line.