Fix logo color in WebKit/Safari
When an SVG image is embedded in a page, WebKit fails to propogate the
prefers-color-sceme adjustment despite the fact that the CSS Working
Group decided that this should be done in secure context. This bug has
been open since 2019, so I'm guessing it won't be fixed anytime soon.

I've decided to cope by inlining the SVG into the HTML document. I've
done this by adding it as a template variable instead of using Pandoc
partials, because I haven't set up the build system to dedect Pandoc
partials as dependencies.

See:
- https://bugs.webkit.org/show_bug.cgi?id=199134
- https://github.com/w3c/csswg-drafts/issues/7213#issuecomment-1183468735