Fix non-compliant timestamps in press kits
The timestamps in the created and published field were being used to
roughly store the time I started working on the project and the time the
project was released, respectively.

This data is interesting, but is not consistent with other timestamps in
my notes. I also don't know how accurate the data is, because when I
looked into some of the dates I found that some of them were wrong. So,
I've decided to drop the information for now.