Add 2014-03-24 21:55 backup
This backup precedes the inception of the git repository and has a more
functional version of the old website, though the blog post data appears
to be more out of date.

The old website used to be hosted on NearlyFreeSpeech, and these files
were stored as a zip file in my backups folder with the name
`2014-03-24 2155.zip`. The time in the name of the backup is likely in
Central Time.

The backup also included some other files; I've only included the ones
related to the old blog. I'm guessing that before the git repository was
made that I was storing all of the website changes using SVN. The SVN
repository has probably been lost to time, as I have been unable to
determine where it is.