Show last commit hash in metadata.
Showing the commit of the current build in the footer means that every
single page has to be rebuilt every time a commit is made. Instead, we
can show show the last commit for the file.

This also makes more sense, because now the commit URL will have
information relevant to the file.