Add favicon
This is the same favicon that used to be on the site when it was still
being generated using Quartz. I was planning on replacing the favicon
with some Haskell code that converted an SVG file into multiple
appropriate favicon files, but it seems like there are no good libraries
for SVG rendering in Haskell. So for now, this will do.