Change main size to 28 cells
While writing a longer blog post, I thought the article was a little
uncomfortable to read due to how wide it was. I had it this wide in
order to accommodate 80 character code blocks, but that's no longer a
concern thanks to horizontal scrolling.

I decided to copy Obsidian; this width more closely resembles the width
of the articles in that program and feels comfortable to read.